feat(it02hu206): added service provider based on HU205

+++ src/app/view/deletePOI/deletePOI.ts
@@ -1,61 +1,41 @@
-import {Component, EventEmitter, Output} from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { POIService } from '../../services/POI/poi.service';
 import { Geohash } from 'geofire-common';
 import { Auth } from '@angular/fire/auth';
+import {POI_REPOSITORY} from '../../services/POI/POIRepository';
+import {POIDB} from '../../services/POI/POIDB';
 
 @Component({
     selector: 'app-delete-confirmation-popup',
     standalone: true,
     imports: [CommonModule],
     templateUrl: './deletePOI.html',
-    styleUrls: ['./deletePOI.css']
+    styleUrls: ['./deletePOI.css'],
+    providers: [POIService, {provide: POI_REPOSITORY, useClass: POIDB}]
 })
 export class DeleteConfirmationPopupComponent {
-    isVisible = true;
-    showSnackbar = false;
+    @Input() geohash: Geohash = " ";
+    // REFACTORIZAR EN IT03, ESTO LO LLEVA AUTHGUARD
+    @Input() auth: Auth | null = null;
     @Output() success = new EventEmitter<boolean>();
-    private snackbarTimeout: any;
-    private geohash: any;
-    private auth: any;
+    @Output() close = new EventEmitter<void>();
 
     constructor(private service: POIService) {}
 
-    // Componente padre abre el popup, inyecta el geohash del POI
-    // ¿Se podría obtener el auth directamente del contexto de la app?
-    open(auth: Auth, geohash: Geohash) {
-        this.isVisible = true;
-        this.geohash = geohash;
-        this.auth = auth;
-    }
-
-    // Cierra el popup
-    onClose() {
-        this.isVisible = false;
-    }
-
     // Ejecuta el borrado (y cierra el popup, si procede)
-    // Propaga el valor obtenido al padre
+    // Propaga el valor obtenido al padre, que es quien muestra el snackbar
+    // Si se ha borrado el POI, debería ser undefined y el padre se cerrará también
     async onConfirm(): Promise<void> {
-        this.success.emit(await this.service.deletePOI(this.auth, this.geohash));
-        this.displaySnackbar();
-    }
-
-    // Mostrar el snackbar
-    private displaySnackbar() {
-        this.showSnackbar = true;
-
-        // Cerrar después de 3 segundos
-        this.snackbarTimeout = setTimeout(() => {
-            this.closeSnackbar();
-        }, 3000);
-    }
-
-    // Cerrar el snackbar
-    closeSnackbar() {
-        this.showSnackbar = false;
-        if (this.snackbarTimeout) {
-            clearTimeout(this.snackbarTimeout);
+        // BORRAR IF EN IT03
+        if (this.auth) {
+            this.success.emit(await this.service.deletePOI(this.auth, this.geohash));
+            this.close.emit();
         }
+    }
+
+    // Envía una señal de cierre
+    onClose() {
+        this.close.emit();
     }
 }
